web-ui/requirements.txt
# Web UI
dash>=2.15.0
dash-bootstrap-components>=1.5.0

# Data handling
influxdb-client>=1.39.0
pandas>=2.2.0
numpy>=1.26.0

# Environment
python-dotenv>=1.0.0
